refactor(validators): move MockAccessor from production to test file
Same cleanup we applied to inputvalidation — test doubles don't belong
in the production binary.
